third_party/WebKit/LayoutTests/bluetooth/device/gattserverdisconnected-event/disconnected.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
 'use strict';
 promise_test(t => {
   return getHealthThermometerDevice()
    .then(({device, fake_peripheral}) => {
      return device.gatt.connect()
        .then(() => {
          // This promise hangs until gattserverdisconnected is fired.
          let disconnected = eventPromise(device, 'gattserverdisconnected');
          fake_peripheral.simulateDisconnect();
          return disconnected;
        })
        .then(event => {
          assert_false(device.gatt.connected);
          assert_true(event.bubbles);
        });
    });
 }, 'A device disconnecting while connected should fire the gattserverdisconnected event.');
</script>
